<food> adapted setrating method, changed tooearly to too_early, added meal ID instead of hashcode, renamed ids into other more specific names <news> updated the thrift file
git-svn-id: file:///var/pocketcampus/trunk@1419 fd477b11-563a-408f-9d6e-eeb9ca009d0c

+++ plugin/food/plugin.food.android/src/org/pocketcampus/plugin/food/android/iface/IFoodModel.java
@@ -86,7 +86,7 @@ public interface IFoodModel {
 	 * Set the <code>Ratings</code> for all meals <code>Meal</code> and notify
 	 * the listeners
 	 */
-	public void setRatings(Map<Integer, Rating> map);
+	public void setRatings(Map<Long, Rating> map);
 
 	/**
 	 * Returns whether the user has already voted or not
